choare(*): update fmea model tests to use class fixture
* Deactivate several tests for modules not yet updated to new API.
