+++ tests/models/programdb/_test_ramstktestmethod.py
@@ -1,15 +1,16 @@
 #!/usr/bin/env python -O
 # -*- coding: utf-8 -*-
 #
-#       tests._dao.programdb.test_ramstkopstress.py is part of The RAMSTK Project
-
+#       tests._dao.programdb.test_ramstktestmethod.py is part of The RAMSTK Project
 #
 # All rights reserved.
-"""Test class for testing the RAMSTKOpStress module algorithms and models."""
+"""Test class for testing the RAMSTKTestMethod module algorithms and models."""
 
+# Third Party Imports
 import pytest
 
-from ramstk.dao.programdb.RAMSTKOpStress import RAMSTKOpStress
+# RAMSTK Package Imports
+from ramstk.dao.programdb.RAMSTKTestMethod import RAMSTKTestMethod
 
 __author__ = 'Doyle Rowland'
 __email__ = 'doyle.rowland@reliaqual.com'
@@ -17,31 +18,29 @@ __organization__ = 'ReliaQual Associates, LLC'
 __copyright__ = 'Copyright 2017 Doyle "weibullguy" Rowland'
 
 ATTRIBUTES = {
-    'description': 'Test Operating Stress',
-    'load_id': 1,
-    'load_history': '',
-    'measurable_parameter': '',
     'remarks': b'',
-    'stress_id': 1
+    'test_id': 1,
+    'boundary_conditions': '',
+    'load_id': 1,
+    'description': 'Test Test Method'
 }
 
 
 @pytest.mark.integration
 def test_ramstkopstress_create(test_dao):
-    """ __init__() should create an RAMSTKOpStress model."""
+    """ __init__() should create an RAMSTKTestMethod model."""
     _session = test_dao.RAMSTK_SESSION(
         bind=test_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKOpStress).first()
+    DUT = _session.query(RAMSTKTestMethod).first()
 
-    assert isinstance(DUT, RAMSTKOpStress)
+    assert isinstance(DUT, RAMSTKTestMethod)
 
     # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_op_stress'
+    assert DUT.__tablename__ == 'ramstk_test_method'
     assert DUT.load_id == 1
-    assert DUT.stress_id == 1
-    assert DUT.description == 'Test Operating Stress'
-    assert DUT.measurable_parameter == ''
-    assert DUT.load_history == ''
+    assert DUT.test_id == 1
+    assert DUT.description == 'Test Test Method'
+    assert DUT.boundary_conditions == ''
     assert DUT.remarks == b''
 
 
@@ -50,17 +49,16 @@ def test_get_attributes(test_dao):
     """ get_attributes() should return a dict of attribute:value pairs. """
     _session = test_dao.RAMSTK_SESSION(
         bind=test_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKOpStress).first()
+    DUT = _session.query(RAMSTKTestMethod).first()
 
     _attributes = DUT.get_attributes()
 
     assert isinstance(_attributes, dict)
 
     assert _attributes['load_id'] == 1
-    assert _attributes['stress_id'] == 1
-    assert _attributes['description'] == 'Test Operating Stress'
-    assert _attributes['load_history'] == ''
-    assert _attributes['measurable_parameter'] == ''
+    assert _attributes['test_id'] == 1
+    assert _attributes['description'] == 'Test Test Method'
+    assert _attributes['boundary_conditions'] == ''
     assert _attributes['remarks'] == b''
 
 
@@ -69,13 +67,13 @@ def test_set_attributes(test_dao):
     """ set_attributes() should return a zero error code on success. """
     _session = test_dao.RAMSTK_SESSION(
         bind=test_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKOpStress).first()
+    DUT = _session.query(RAMSTKTestMethod).first()
 
     _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
 
     assert _error_code == 0
-    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKOpStress {0:d} "
-                    "attributes.".format(DUT.stress_id))
+    assert _msg == ("RAMSTK SUCCESS: Updating RAMSTKTestMethod {0:d} "
+                    "attributes.".format(DUT.load_id))
 
 
 @pytest.mark.integration
@@ -83,7 +81,7 @@ def test_set_attributes_missing_key(test_dao):
     """ set_attributes() should return a 40 error code when passed a dict with a missing key."""
     _session = test_dao.RAMSTK_SESSION(
         bind=test_dao.engine, autoflush=False, expire_on_commit=False)
-    DUT = _session.query(RAMSTKOpStress).first()
+    DUT = _session.query(RAMSTKTestMethod).first()
 
     ATTRIBUTES.pop('description')
 
@@ -91,6 +89,6 @@ def test_set_attributes_missing_key(test_dao):
 
     assert _error_code == 40
     assert _msg == ("RAMSTK ERROR: Missing attribute 'description' in attribute "
-                    "dictionary passed to RAMSTKOpStress.set_attributes().")
+                    "dictionary passed to RAMSTKTestMethod.set_attributes().")
 
-    ATTRIBUTES['description'] = 'Test Operating Stress'
+    ATTRIBUTES['description'] = 'Test Test Method'
