+++ tests/models/programdb/test_ramstkaction.py
@@ -17,12 +17,12 @@ import pytest
 from ramstk.models.programdb.RAMSTKAction import RAMSTKAction
 
 ATTRIBUTES = {
-    'action_due_date': date.today() + timedelta(days=30),
-    'action_approve_date': date.today() + timedelta(days=30),
+    'action_due_date': date(2019, 8, 20),
+    'action_approve_date': date(2019, 8, 20),
     'action_status': '',
     'action_closed': 0,
     'action_taken': b'',
-    'action_close_date': date.today() + timedelta(days=30),
+    'action_close_date': date(2019, 8, 20),
     'action_recommended': b'Recommended action for Failure Cause #1',
     'action_category': '',
     'action_owner': '',
@@ -30,67 +30,66 @@ ATTRIBUTES = {
 }
 
 
-@pytest.mark.integration
-def test_ramstkaction_create(test_dao):
-    """__init__() should create an RAMSTKAction model."""
-    DUT = test_dao.session.query(RAMSTKAction).first()
-    assert isinstance(DUT, RAMSTKAction)
+@pytest.mark.usefixtures('test_program_dao')
+class TestRAMSTKAction():
+    """Class for testing the RAMSTKAction model."""
+    @pytest.mark.integration
+    def test_ramstkaction_create(self,  test_program_dao):
+        """__init__() should create an RAMSTKAction model."""
+        DUT = test_program_dao.session.query(RAMSTKAction).first()
+        assert isinstance(DUT, RAMSTKAction)
 
-    # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_action'
-    assert DUT.cause_id == 1
-    assert DUT.action_id == 1
-    assert DUT.action_recommended == (
-        b'Test Functional FMEA Recommended Action #1 for Cause ID 1'
-    )
-    assert DUT.action_category == ''
-    assert DUT.action_owner == ''
-    assert DUT.action_due_date == date.today() + timedelta(days=30)
-    assert DUT.action_status == ''
-    assert DUT.action_taken == b''
-    assert DUT.action_approved == 0
-    assert DUT.action_approve_date == date.today() + timedelta(days=30)
-    assert DUT.action_closed == 0
-    assert DUT.action_close_date == date.today() + timedelta(days=30)
-
-
-@pytest.mark.integration
-def test_get_attributes(test_dao):
-    """
-    get_attributes() should return a dict of attribute:value pairs.
-    """
-    DUT = test_dao.session.query(RAMSTKAction).first()
-
-    _attributes = DUT.get_attributes()
-
-    assert isinstance(_attributes, dict)
+        # Verify class attributes are properly initialized.
+        assert DUT.__tablename__ == 'ramstk_action'
+        assert DUT.cause_id == 1
+        assert DUT.action_id == 1
+        assert DUT.action_recommended == (
+            b'Recommended action for Failure Cause #1'
+        )
+        assert DUT.action_category == ''
+        assert DUT.action_owner == ''
+        assert DUT.action_due_date == date(2019, 8, 20)
+        assert DUT.action_status == ''
+        assert DUT.action_taken == b''
+        assert DUT.action_approved == 0
+        assert DUT.action_approve_date == date(2019, 8, 20)
+        assert DUT.action_closed == 0
+        assert DUT.action_close_date == date(2019, 8, 20)
 
 
-@pytest.mark.integration
-def test_set_attributes(test_dao):
-    """
-    set_attributes() should return a zero error code on success.
-    """
-    DUT = test_dao.session.query(RAMSTKAction).first()
+    @pytest.mark.integration
+    def test_get_attributes(self, test_program_dao):
+        """get_attributes() should return a dict of attribute:value pairs."""
+        DUT = test_program_dao.session.query(RAMSTKAction).first()
 
-    assert DUT.set_attributes(ATTRIBUTES) is None
+        _attributes = DUT.get_attributes()
 
-
-@pytest.mark.integration
-def test_set_attributes_none_value(test_dao):
-    """set_attributes() should set an attribute to it's default value when the attribute is passed with a None value."""
-    DUT = test_dao.session.query(RAMSTKAction).first()
-
-    ATTRIBUTES['action_status'] = None
-
-    assert DUT.set_attributes(ATTRIBUTES) is None
-    assert DUT.get_attributes()['action_status'] == ''
+        assert isinstance(_attributes, dict)
 
 
-@pytest.mark.integration
-def test_set_attributes_unknown_attributes(test_dao):
-    """set_attributes() should raise an AttributeError when passed an unknown attribute."""
-    DUT = test_dao.session.query(RAMSTKAction).first()
+    @pytest.mark.integration
+    def test_set_attributes(self, test_program_dao):
+        """set_attributes() should return a zero error code on success."""
+        DUT = test_program_dao.session.query(RAMSTKAction).first()
 
-    with pytest.raises(AttributeError):
-        DUT.set_attributes({'shibboly-bibbly-boo': 0.9998})
+        assert DUT.set_attributes(ATTRIBUTES) is None
+
+
+    @pytest.mark.integration
+    def test_set_attributes_none_value(self, test_program_dao):
+        """set_attributes() should set an attribute to it's default value when the attribute is passed with a None value."""
+        DUT = test_program_dao.session.query(RAMSTKAction).first()
+
+        ATTRIBUTES['action_status'] = None
+
+        assert DUT.set_attributes(ATTRIBUTES) is None
+        assert DUT.get_attributes()['action_status'] == ''
+
+
+    @pytest.mark.integration
+    def test_set_attributes_unknown_attributes(self, test_program_dao):
+        """set_attributes() should raise an AttributeError when passed an unknown attribute."""
+        DUT = test_program_dao.session.query(RAMSTKAction).first()
+
+        with pytest.raises(AttributeError):
+            DUT.set_attributes({'shibboly-bibbly-boo': 0.9998})
