+++ tests/models/programdb/test_ramstkcause.py
@@ -24,70 +24,69 @@ ATTRIBUTES = {
 }
 
 
-@pytest.mark.integration
-def test_ramstkcause_create(test_dao):
-    """ __init__() should create an RAMSTKCause model. """
-    DUT = test_dao.session.query(RAMSTKCause).first()
+@pytest.mark.usefixtures('test_program_dao')
+class TestRAMSTKCause():
+    """Class for testing the RAMSTKCause model."""
+    @pytest.mark.integration
+    def test_ramstkcause_create(self, test_program_dao):
+        """ __init__() should create an RAMSTKCause model. """
+        DUT = test_program_dao.session.query(RAMSTKCause).first()
 
-    assert isinstance(DUT, RAMSTKCause)
+        assert isinstance(DUT, RAMSTKCause)
 
-    # Verify class attributes are properly initialized.
-    assert DUT.__tablename__ == 'ramstk_cause'
-    assert DUT.mode_id == 1
-    assert DUT.cause_id == 1
-    assert DUT.description == 'Test Functional FMEA Cause #1 for Mode ID 1'
-    assert DUT.rpn == 0
-    assert DUT.rpn_detection == 0
-    assert DUT.rpn_detection_new == 0
-    assert DUT.rpn_new == 0
-    assert DUT.rpn_occurrence == 0
-    assert DUT.rpn_occurrence_new == 0
+        # Verify class attributes are properly initialized.
+        assert DUT.__tablename__ == 'ramstk_cause'
+        assert DUT.mode_id == 1
+        assert DUT.cause_id == 1
+        assert DUT.description == 'Test Functional FMEA Cause #1 for Mode ID 1'
+        assert DUT.rpn == 0
+        assert DUT.rpn_detection == 0
+        assert DUT.rpn_detection_new == 0
+        assert DUT.rpn_new == 0
+        assert DUT.rpn_occurrence == 0
+        assert DUT.rpn_occurrence_new == 0
 
+    @pytest.mark.integration
+    def test_get_attributes(self, test_program_dao):
+        """ get_attributes() should return a tuple of attribute values. """
+        DUT = test_program_dao.session.query(RAMSTKCause).first()
 
-@pytest.mark.integration
-def test_get_attributes(test_dao):
-    """ get_attributes() should return a tuple of attribute values. """
-    DUT = test_dao.session.query(RAMSTKCause).first()
+        _attributes = DUT.get_attributes()
 
-    _attributes = DUT.get_attributes()
+        assert isinstance(_attributes, dict)
+        assert _attributes['mode_id'] == 1
+        assert _attributes['mechanism_id'] == -1
+        assert _attributes['cause_id'] == 1
+        assert _attributes['description'] == ('Test Functional FMEA Cause #1 for '
+                                              'Mode ID 1')
+        assert _attributes['rpn'] == 0
+        assert _attributes['rpn_detection'] == 0
+        assert _attributes['rpn_detection_new'] == 0
+        assert _attributes['rpn_new'] == 0
+        assert _attributes['rpn_occurrence'] == 0
+        assert _attributes['rpn_occurrence_new'] == 0
 
-    assert isinstance(_attributes, dict)
-    assert _attributes['mode_id'] == 1
-    assert _attributes['mechanism_id'] == -1
-    assert _attributes['cause_id'] == 1
-    assert _attributes['description'] == ('Test Functional FMEA Cause #1 for '
-                                          'Mode ID 1')
-    assert _attributes['rpn'] == 0
-    assert _attributes['rpn_detection'] == 0
-    assert _attributes['rpn_detection_new'] == 0
-    assert _attributes['rpn_new'] == 0
-    assert _attributes['rpn_occurrence'] == 0
-    assert _attributes['rpn_occurrence_new'] == 0
+    @pytest.mark.integration
+    def test_set_attributes(self, test_program_dao):
+        """ set_attributes() should return a zero error code on success. """
+        DUT = test_program_dao.session.query(RAMSTKCause).first()
 
+        assert DUT.set_attributes(ATTRIBUTES) is None
 
-@pytest.mark.integration
-def test_set_attributes(test_dao):
-    """ set_attributes() should return a zero error code on success. """
-    DUT = test_dao.session.query(RAMSTKCause).first()
+    @pytest.mark.integration
+    def test_set_attributes_none_value(self, test_program_dao):
+        """set_attributes() should set an attribute to it's default value when the attribute is passed with a None value."""
+        DUT = test_program_dao.session.query(RAMSTKCause).first()
 
-    assert DUT.set_attributes(ATTRIBUTES) is None
+        ATTRIBUTES['description'] = None
 
+        assert DUT.set_attributes(ATTRIBUTES) is None
+        assert DUT.get_attributes()['description'] == ''
 
-@pytest.mark.integration
-def test_set_attributes_none_value(test_dao):
-    """set_attributes() should set an attribute to it's default value when the attribute is passed with a None value."""
-    DUT = test_dao.session.query(RAMSTKCause).first()
+    @pytest.mark.integration
+    def test_set_attributes_unknown_attributes(self, test_program_dao):
+        """set_attributes() should raise an AttributeError when passed an unknown attribute."""
+        DUT = test_program_dao.session.query(RAMSTKCause).first()
 
-    ATTRIBUTES['description'] = None
-
-    assert DUT.set_attributes(ATTRIBUTES) is None
-    assert DUT.get_attributes()['description'] == ''
-
-
-@pytest.mark.integration
-def test_set_attributes_unknown_attributes(test_dao):
-    """set_attributes() should raise an AttributeError when passed an unknown attribute."""
-    DUT = test_dao.session.query(RAMSTKCause).first()
-
-    with pytest.raises(AttributeError):
-        DUT.set_attributes({'shibboly-bibbly-boo': 0.9998})
+        with pytest.raises(AttributeError):
+            DUT.set_attributes({'shibboly-bibbly-boo': 0.9998})
